(PC-34542) feat(chronicle): remark
