+++ src/features/chronicle/components/ChronicleCardListHeader/ChronicleCardListHeader.tsx
@@ -1,28 +1,25 @@
 import React, { FunctionComponent } from 'react'
 
-import { ChroniclesWritersModal } from 'features/chronicle/pages/ChroniclesWritersModal/ChroniclesWritersModal'
 import { ButtonQuaternaryBlack } from 'ui/components/buttons/ButtonQuaternaryBlack'
 import { styledButton } from 'ui/components/buttons/styledButton'
-import { useModal } from 'ui/components/modals/useModal'
 import { ViewGap } from 'ui/components/ViewGap/ViewGap'
 import { InfoPlain } from 'ui/svg/icons/InfoPlain'
 import { TypoDS, getSpacing } from 'ui/theme'
 
-export const ChronicleCardListHeader: FunctionComponent = () => {
-  const { visible, showModal, hideModal } = useModal(false)
+type Props = {
+  onPressMoreInfo: VoidFunction
+}
 
+export const ChronicleCardListHeader: FunctionComponent<Props> = ({ onPressMoreInfo }) => {
   return (
-    <React.Fragment>
-      <ViewGap gap={2}>
-        <TypoDS.Title2>Tous les avis</TypoDS.Title2>
-        <StyledButtonQuaternaryBlack
-          wording="Qui écrit les avis&nbsp;?"
-          icon={InfoPlain}
-          onPress={showModal}
-        />
-      </ViewGap>
-      <ChroniclesWritersModal closeModal={hideModal} isVisible={visible} />
-    </React.Fragment>
+    <ViewGap gap={2}>
+      <TypoDS.Title2>Tous les avis</TypoDS.Title2>
+      <StyledButtonQuaternaryBlack
+        wording="Qui écrit les avis&nbsp;?"
+        icon={InfoPlain}
+        onPress={onPressMoreInfo}
+      />
+    </ViewGap>
   )
 }
