+++ src/features/chronicle/pages/Chronicles/ChroniclesBase.tsx
@@ -9,9 +9,11 @@ import { ChronicleCardList } from 'features/chronicle/components/ChronicleCardLi
 import { ChronicleCardListHeader } from 'features/chronicle/components/ChronicleCardListHeader/ChronicleCardListHeader'
 import { ChroniclesHeader } from 'features/chronicle/components/ChroniclesHeader/ChroniclesHeader'
 import { ChroniclesWebMetaHeader } from 'features/chronicle/components/ChroniclesWebMetaHeader/ChroniclesWebMetaHeader'
+import { ChroniclesWritersModal } from 'features/chronicle/pages/ChroniclesWritersModal/ChroniclesWritersModal'
 import { ChronicleCardData } from 'features/chronicle/type'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
+import { useModal } from 'ui/components/modals/useModal'
 import { getSpacing } from 'ui/theme'
 
 type Props = PropsWithChildren<{
@@ -34,6 +36,7 @@ export const ChroniclesBase: FunctionComponent<Props> = ({
   const headerHeight = appBarHeight + top
 
   const { headerTransition, onScroll } = useOpacityTransition()
+  const { visible, showModal, hideModal } = useModal(false)
 
   const chroniclesListRef = useRef<FlatList<ChronicleCardData>>(null)
 
@@ -72,7 +75,7 @@ export const ChroniclesBase: FunctionComponent<Props> = ({
           data={chronicleCardsData}
           horizontal={false}
           separatorSize={6}
-          headerComponent={<ChronicleCardListHeader />}
+          headerComponent={<ChronicleCardListHeader onPressMoreInfo={showModal} />}
           ref={chroniclesListRef}
           onScroll={onScroll}
           contentContainerStyle={{
@@ -88,6 +91,7 @@ export const ChroniclesBase: FunctionComponent<Props> = ({
           onLayout={handleLayout}
         />
       </FullFlexRow>
+      <ChroniclesWritersModal closeModal={hideModal} isVisible={visible} />
     </React.Fragment>
   )
 }
